+++ Plugins/PluginUtilities/PluginUtilities.vcxitems
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup Label="Globals">
     <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
@@ -15,8 +15,10 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="$(MSBuildThisFileDirectory)PluginUtilities.cpp" />
+    <ClCompile Include="$(MSBuildThisFileDirectory)ZoneUtilities.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="$(MSBuildThisFileDirectory)PluginUtilities.h" />
+    <ClInclude Include="$(MSBuildThisFileDirectory)ZoneUtilities.h" />
   </ItemGroup>
 </Project>
